New module for 'assert_meeting_code'; 'attend_class' is now a package; Tests folder added, yet not implemented;
	new file:   src/attendClass/__init__.py
	new file:   src/attendClass/assert_meeting_code.py
	renamed:    src/attendClass.py -> src/attendClass/attendClass.py
	new file:   tests/attendTest.py
	new file:   tests/openNewTest.py
