+++ src/attendClass/__init__.py
@@ -1,0 +1,2 @@
+from .attendClass import AttendClass
+from .assert_meeting_code import assert_meeting_code
