+++ tests/openNewTest.py
@@ -1,0 +1,1 @@
+raise NotImplementedError
